Updated Source  + Jar

+++ Source Code/Sparkz Enginev0.08 - 0.8/src/com/engine/ParticleTypes/BlackHole.java
@@ -9,7 +9,7 @@ public class BlackHole extends Molecule {
     public BlackHole() {super();}
 
     public BlackHole(double _x, double _y, double _radius, double speed, int direction) {
-        super(_x, _y, _radius, speed, direction, (byte) 0);
+        super(_x, _y, _radius, speed, direction, 0);
     }
 
     private void open() {
