+++ Source Code/Sparkz Enginev0.08 - 0.8/src/com/engine/ParticleTypes/Eraser.java
@@ -10,7 +10,7 @@ public class Eraser extends Molecule {
 
     public Eraser(){super();}
     public Eraser(double _x, double _y, double _radius, double speed, int direction) {
-        super(_x, _y, _radius, speed, direction, (byte) 0);
+        super(_x, _y, _radius, speed, direction, 0);
         color = Color.orange;
     }
