+++ Source Code/Sparkz Enginev0.08 - 0.8/src/com/engine/ParticleTypes/Fireworks.java
@@ -11,11 +11,11 @@ public class Fireworks extends Molecule {
 
     public Fireworks(){super();}
     public Fireworks(double _x, double _y, double _radius, double speed, int direction) {
-        super(_x, _y, _radius, speed, direction, (byte) 0);
+        super(_x, _y, _radius, speed, direction, 0);
     }
 
     public Fireworks(double _x, double _y, double _radius, double speed, int direction, int windval) {
-        super(_x, _y, _radius, speed, direction, (byte) 0);
+        super(_x, _y, _radius, speed, direction, 0);
         wind = windval;
     }
