+++ Source Code/Sparkz Enginev0.08 - 0.8/src/com/engine/ParticleTypes/Flux.java
@@ -10,7 +10,7 @@ public class Flux extends Molecule {
 
     public Flux(){super();}
     public Flux(double _x, double _y, double _radius, double speed, int direction) {
-        super(_x, _y, _radius, speed, direction, (byte) 0);
+        super(_x, _y, _radius, speed, direction, 0);
     }
 
     public void giveStyle() {
