+++ Source Code/Sparkz Enginev0.08 - 0.8/src/com/engine/ParticleTypes/Ion.java
@@ -8,7 +8,7 @@ public class Ion extends Molecule {
     public Ion() {super();}
 
     public Ion(double _x, double _y, double _radius, double speed, int direction) {
-        super(_x, _y, _radius, speed, direction, (byte) 0);
+        super(_x, _y, _radius, speed, direction, 0);
     }
 
     private void coagulate() {
